HZN-1345: Move classes to root package org.opennms.core.health
