+++ core/health/shell/pom.xml
@@ -22,8 +22,7 @@
                         <Bundle-SymbolicName>${project.groupId}.${project.artifactId}</Bundle-SymbolicName>
                         <Bundle-Version>${project.version}</Bundle-Version>
                         <Private-Package></Private-Package>
-                        <!-- TODO MVR change package name -->
-                        <Karaf-Commands>org.opennms.distributed.core.health.*</Karaf-Commands>
+                        <Karaf-Commands>org.opennms.core.health.*</Karaf-Commands>
                     </instructions>
                 </configuration>
             </plugin>
