+++ features/distributed/core/impl/src/main/java/org/opennms/distributed/core/impl/RestConnectionHealthCheck.java
@@ -29,10 +29,10 @@
 package org.opennms.distributed.core.impl;
 
 import org.opennms.distributed.core.api.RestClient;
-import org.opennms.distributed.core.health.Context;
-import org.opennms.distributed.core.health.HealthCheck;
-import org.opennms.distributed.core.health.Response;
-import org.opennms.distributed.core.health.Status;
+import org.opennms.core.health.api.Context;
+import org.opennms.core.health.api.HealthCheck;
+import org.opennms.core.health.api.Response;
+import org.opennms.core.health.api.Status;
 
 public class RestConnectionHealthCheck implements HealthCheck {
