+++ features/distributed/datasource-healthcheck/src/main/java/org/opennms/netmgt/distributed/datasource/healthcheck/DatasourceHealthCheck.java
@@ -30,10 +30,10 @@ package org.opennms.netmgt.distributed.datasource.healthcheck;
 
 import java.util.Objects;
 
-import org.opennms.distributed.core.health.Context;
-import org.opennms.distributed.core.health.HealthCheck;
-import org.opennms.distributed.core.health.Response;
-import org.opennms.distributed.core.health.Status;
+import org.opennms.core.health.api.Context;
+import org.opennms.core.health.api.HealthCheck;
+import org.opennms.core.health.api.Response;
+import org.opennms.core.health.api.Status;
 import org.osgi.framework.BundleContext;
 import org.osgi.framework.ServiceReference;
