+++ features/distributed/datasource-healthcheck/src/main/resources/OSGI-INF/blueprint/blueprint.xml
@@ -13,7 +13,7 @@
 		http://aries.apache.org/schemas/blueprint-ext/blueprint-ext-1.5.xsd
 ">
 
-    <service interface="org.opennms.distributed.core.health.HealthCheck">
+    <service interface="org.opennms.core.health.api.HealthCheck">
         <bean class="org.opennms.netmgt.distributed.datasource.healthcheck.DatasourceHealthCheck" >
             <argument ref="blueprintBundleContext"/>
         </bean>
